clean:カスタムフックをclientとserverに分離 #33
